coredump: allow systemd-coredump to run during shutdown
Ever since systemd-coredump was made a socket-activated service in
systemd 229, it has been unable to run during system shutdown.

This is a problem if any processes crash during the shutdown sequence.
If a process crashes when it receives SIGTERM, then the kernel tries to
send the coredump to systemd-coredump.  systemd-coredump passes the
coredump file descriptor onto the socket.  However, systemd is then
unable to start systemd-coredump@.service due to that unit conflicting
shutdown.target.  This causes the file descriptor to be kept open, making
the crashing process stay alive and unresponsive to SIGTERM.  systemd
then must wait its default timeout of 90 seconds before sending SIGKILL,
significantly delaying the shutdown.

This is not a hypothetical case; I initially encountered it with the
chromium browser.  systemd cannot assume that applications do not crash.

Fix the problem by making the following changes:

- Define a new slice coredump.slice with DefaultDependencies=No and place
  systemd-coredump@.service in it.  This is necessary because by default
  templated services are placed in a per-template slice which conflicts
  shutdown.target, thereby causing them to be stopped at shutdown.  Note
  that because the '-' character is special in slice names, I could not
  call the new slice systemd-coredump.slice.  Also, an alternative would
  have been to run the coredump services in system.slice; however, this
  was recommended against because coredump handling can be expensive.

- Remove Conflicts=shutdown.target and Before=shutdown.target from
  systemd-remount-fs.service.  This prevents this early boot service,
  which is ordered before systemd-coredump@.service, from being stopped
  at shutdown.  Stopping this service was a no-op anyway, since it is
  oneshot.

- Finally, remove Conflicts=shutdown.target and Before=shutdown.target
  from systemd-coredump@.service.  In combination with the other changes,
  this allows systemd-coredump@.service to be started during shutdown.

Fixes #2691.