install: cache the presets before evaluating
The previous implementation traversed the various config directories,
walking the preset files and parsing each line to determine if a service
should be enabled or disabled. It did this for every service which
resulted in a combinatorial explosion.

This approach parses each of the preset entries into a hashmap which is
then used to check if each service should be enabled or disabled.