cgtop: initialize `ours' to NULL propery
Running cgtop on a system, which lacks expecting stat file, results to
segfault. For example, a system with blkio tree but without cfq io
scheduler, lacks "blkio.io_service_bytes".

When the targeting cgroup's file does not exist, process() return 0 and
also does not modify `*ret' value (which is `*ours'). As a result,
callers of refresh_one() can have bogus pointer, which result to SEGV.

This patch just propery initialize the variable to NULL.