fstab-generator: add support for the nfs bg mountoption
the nfs manpage shows:

-->--
If the bg option is specified, a timeout or failure causes the mount(8) command
to fork a child which continues to attempt to mount the export.  The parent
immediately returns with a zero exit code.  This is known as a "background"
mount.
--<--

This looks similar to the nofail mount option.