rules: call btrfs ready with the /dev/mapper/name after the link is created
Most Linux file systems just report the name of the device node used
to mount the file system in /proc/mounts.  Mount(8) canonicalizes
device mapper names to use the "friendly" /dev/mapper/name nodes before
calling the mount system call.  That translates directly into tools like
'mount' and 'df' displaying the friendly names without any additional
processing.

Btrfs is different.  Since it supports the ability to span multiple
devices and the device used to mount the file system initially can be
removed, it needs to report the device name for /proc/mounts itself.
Btrfs knows about the block devices it can use but only has names for
them as passed from userspace.  The "btrfs device ready" command in the
udev rule tells btrfs about the block devices but uses the $devnode name
for it since that's all that can be assumed to exist at that time.  When
the file system is mounted using the friendly name, btrfs adopts that
name for the block device.  If that device is removed, btrfs still
needs to provide a name and all that's left are the device nodes
passed as $devnode, so we get /dev/dm-# instead.  We've had customers
report this issue and I've reproduced it myself.

We can resolve the issue by adding a RUN rule that calls
"btrfs device ready" again after the symlink has been created so that
the device list has the friendly names.  This is safe for btrfs since
it doesn't actually care what the names are other than for passing
them back to the user via /proc/mounts.