Add missing search paths in modules-load.d manpage
I was debugging a failing systemd-modules-load service and found the offending line in /lib/modules-load.d/ which was not mentioned in the man page. While the /usr/local/lib directory is searched unconditionally, the /lib/ directory depends on the HAVE_SPLIT_USR flag. I am not sure if an according note would be necessary in addition.