networkd: Do not start ndisc client if link is a bridge port
We should not start ndisc client when a link is bridge port.

Bridge could be configured by us.
by looking at conf files we know what this link is a bridge
port and should not start ndisc.

Bridge already exists with ports. (nothing in networkd conf)
Here we need to find from the netlink messages if anylink
is bridge port.

bridge can be created by external tool and ports can be added
to it. netlink messages will inform us and we won't start ndisc
client.

To know link is bridge port :
If a link enslaved to a bridge kernel gives us IFLA_PROTINFO and
IFLA_BRPORT_STATE, IFLA_BRPORT_PRIORITY etc. We just have to look is
there attributes are present. and determine if this link belongs to a
bridge.

fixes #2572

Note: If a interface already aquired addresses we dont drop those.
This will be a future RFE.