systemd-fsck@ doesn't support labels with slashes in them (#3300)
Even though it shows the right path, it doesn't encode the label properly.
This fix is pretty ugly, as the right one would be to encode the path properly, but this is way too intricated to risk this.

Instead, check if a mount by label fail, try to "fix" the label, and see if it succeeds.

Tested it on a host with a /data mounted by label, and it worked.

Fixes #3300